requirements.txt
django==3.2.13
djangorestframework==3.14.0
model-bakery==1.15.0
pytest==7.4.2
pytest-django==4.5.2
python-decouple==3.8
psycopg2==2.9.7
